property_definitions: split to_boolean
All of the other property functions have a static version
(to_double, to_long, etc), and a class method that finds the item
and then passes on to the static to_xxx version.

This exists so we can reuse those functions in places where we don't
have a fully constructed property_definitions map. It will happen if
we are parsing sub-options, in compaction strategy for instance.

Currently we're not doing that with to_boolean. This patch adds that
to match the other properties.

Signed-off-by: Glauber Costa <glauber@scylladb.com>