STCS: merge buckets if none of them is interesting
get_buckets will return a vector of buckets ordered by their average
size. If none of them is currently interesting according to the
min_threshold criteria we should try to merge buckets and see if that
makes a compaction become interesting. Essentially what this does is to
allow cross-tier compaction.

Because the buckets are sorted by their average SSTable-size, we will walk
them in order and try to find two SSTables that are close enough to each
other, so that the ratio between their sizes is the minimum possible.

Signed-off-by: Glauber Costa <glauber@scylladb.com>