STCS: merge upper level buckets if none of them is interesting
get_buckets will return a vector of buckets ordered by their average
size. If none of them is currently interesting according to the
min_threshold criteria we should try to merge buckets and see if that
makes a compaction become interesting. Essentially what this does is to
allow cross-tier compaction.

Because the buckets are sorted by their average SSTable-size, we want to
do that in the last bucket.

Signed-off-by: Glauber Costa <glauber@scylladb.com>