sstables: let compaction strategy choose which directory to write to
Right now there is only a single directory hosting SSTables. But we may have
more than one in the future. When we are compacting a new SSTable, we should
allow the compactor process to choose in which directory we should be writing
SSTables. The compactor will have discretion on how to choose this, but right
now we only pass the current single-directory as a parameter.

Signed-off-by: Glauber Costa <glauber@scylladb.com>