database: make nodetool listsnapshots print correct information
nodetool listsnapshots is not working at all, and is particularly broken
with la files.

The first reason is easy: we are moving the snapshot directory name
before we use it. But even if we fix that, it still doesn't work for la files.

The reason is that we validate if the file is really an SSTable - which is a
good thing, but the la regex says this:

   static std::regex dir(".*/([^/]*)/(\\w+)-[\\da-fA-F]+(?:/upload)?/?");

But the snapshot paths have a snapshot part that the regex does not
recognize. Since the regex is already a bit complex, to avoid making it
even more complex we can note that the only reason we are passing the
file through the regex is to make sure the file itself is not bogus and
really refers to an SSTable. We will not open it. So we can just pass
the original datadir instead.

Fixes #3572

Signed-off-by: Glauber Costa <glauber@scylladb.com>
---
v2: mention correct bug number