database: generate compaction_history entries for all compaction types.
We have recently demoted our compaction messages in the log to DEBUG,
meaning users have to rely on the compaction_history table to learn
information about their compactions.

The problem is that historically we've only been adding normal automatic
compactions to that table. If we are to really have a replacement to
the log messages, we need to log them all.

Signed-off-by: Glauber Costa <glauber@scylladb.com>