compaction: augment compaction info
There is some information that we would like to add to the compaction
history table, that is simply not available on the compaction_info
object.

In particular, that is the time in which the compaction started-- we
just record the time in which it ended, and the list of old SSTables--
we just record the list of SSTables generated.

The list of old SSTables can actually be acquired by other means around
the point in which we need to use it, but having it in the
compaction_info will simplify usage a lot and it shouldn't create a huge
expense.

Signed-off-by: Glauber Costa <glauber@scylladb.com>