sstables: add type, start type, and sstable lists to compaction history
We have recently demoted the compaction messages to DEBUG, in the hopes
that people will check the compaction_history table when needed.
However, compaction_history lacks important values that we have used
many times to figure out potential issues.

Those are, the type of compaction (resharding, compaction, cleanup,
etc), the time compaction started (so we can calculate deltas with
the time it ended), and the list of SSTables involved.

About that list, it would probably be more compact to use just the generations.
I still opted to use the filename, though, since this can help us figure out
issues with compactions that are using files from older formats and
outputting files into new formats.

Remarks:

- there is no need to worry about altering table for the
compaction_history table, because system table schemas are in memory. I
have tested that by booting a version with this patch over a directory
created with master, and everything works.

Fixes #3791

Signed-off-by: Glauber Costa <glauber@scylladb.com>