relocatable-package: do not package the entire dist directory
We are packaging the entire dist directory, but that directory
contains AMI files, redhat and debian build files and files for
the offline installer aside from the Scylla scripts.

We only need the Scylla scripts to make it into the package, so change
the relocatable process to include just that.

In a follow up patch, I will transform the scripts in that directory
before inclusion.

Signed-off-by: Glauber Costa <glauber@scylladb.com>