install-dependencies.sh: add packages that will be needed by scylla-python3
Done in a separate step so we can update the toolchain first.

dnf-utils is used to bring us repoquery, which we will use to derive the
list of files in the python packages.
patchelf is needed so we can add a DT_RUNPATH section to the interpreter
binary.
the python modules, as well as the python3 interpreter are taken from
the current RPM spec file.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
[avi: regenerate frozen toolchain image]
Message-Id: <20190123011751.14440-1-glauber@scylladb.com>
(cherry picked from commit 5d754c1d1114ed59ccb41b8bdbe4ea0c05cb76ae)
Signed-off-by: Glauber Costa <glauber@scylladb.com>

Conflicts:
	tools/toolchain/image

(No toolchain in 3.0)