reshard synchronously instead of asynchronously
This patch changes resharding so that it returns synchronously instead
of asynchronously. load_new_sstables will now only return after
resharding, which is okay for an import operation. In subsequent patches
we will change load_new_sstables even further so it is fully
offstrategy-friendly.

The biggest change is when resharding happens upon node restart. Whereas
before this patch we would start resharding but bring the node online as
soon as possible, we will now wait for resharding to complete before
moving on.

Resharding is faster this way since it doesn't have to compete with
anything else: not only there are no reads or writes at this point, we
also disable automatic compactions. That means that the node will be
able to use its full resources towards resharding. When reads finally
start, we will greatly reduce the number of SSTables needed to serve
a request, since shared SSTables will not exist.

Note how the commitlog is also only replayed after resharding is done,
so keeping automatic compactions disabled is safe.

By simply moving code around, resharding becomes non-interruptible:
this is because the stop_signal is only checked once all services are
initialized. To aid with that, we will register a new listener on
the abort_source: it will stop the compaction manager which as a side
effect will cancel all existing reshard operations.

Tests:
- unit tests (dev)
- resharding dtest.
- manual reshards
- manual nodetool refresh
- ctrl+c the node during reshard
- ctrl+c the node during nodetool refresh

Signed-off-by: Glauber Costa <glauber@scylladb.com>