compaction: pass descriptor all the way down to compaction object.
To do that - and still avoid a copy - we need to add some fields
to the compaction object that are exclusive to regular_compaction.
Still, not only this simplifies the code, resharding and regular
compaction look more and more alike.

This is done now in preparation for another patch that will add
more information to the descriptor.

Signed-off-by: Glauber Costa <glauber@scylladb.com>