compaction: make the sstable_writer optional
The end game here is to introduce compactions that may not write
anything.  Instead, they just scan all the data. We will do that with
the SSTables that currently sit in staging.

Reusing the compaction mechanism has advantages on its own, but the
main advantage is that in a subsequent step we can combine both: a
compaction that can optionally generate streaming data, including
view generations, and write SSTables. This will be useful to handle
SSTables dumped in the upload directory in a single pass.

Signed-off-by: Glauber Costa <glauber@scylladb.com>