compaction: use the more standard abort source to stop a compaction
Will be handy when we try to integrate compaction with view_building,
which already uses abort sources to stop itself.

Compactions are stopped from the outside, by the compaction_manager: the
compaction_manager only has access to the compaction_info.

The abort_source could sit in the compaction_info itself, but we need
compaction_info to be noexcept move-constructible since it is moved
around and it is not. Therefore we create the object in the compaction
object and pass a pointer to the compaction_info

Note that we need one abort source per compaction, since we not always
stop all compactions together. We can stop compactions on a table basis,
and soon, hopefully, on an individual basis.

Signed-off-by: Glauber Costa <glauber@scylladb.com>