distributed_loader: reshard before the node is made online
This patch moves the resharding process to use the new
directory_with_sstables_handler infrastructure. There is no longer
a clear reshard step, and that just becomes a natural part of
populate_column_family.

In main.cc, a couple of changes are necessary to make that happen.
The first one obviously is to stop calling reshard. We also need to
make sure that:
 - The compaction manager is started much earlier, so we can register
   resharding jobs with it.
 - resharding, which can take a long time, can be aborted. So we
   start the commitlog_manager with an abort_source
 - auto compactions are disabled in the populate method, so resharding
   doesn't have to fight for bandwidth with auto compactions. Note that
   in the end game of this we'll do an extra round in which we make those
   SSTables in-strategy. main.cc has to enable them when all potential
   resharding acitvity is done (after non system keyspaces are started)

Signed-off-by: Glauber Costa <glauber@scylladb.com>