compaction_manager: allow early aborts through abort sources.
The shutdown process of compaction manager starts with an explicit call
from the database object. However that can only happen everything is
already initialized. This works well today, but I am soon to change
the resharding process to operate before the node is fully ready.

One can still stop the database in this case, but reshardings will
have to finish before the abort signal is processed.

This patch passes the existing abort source to the construction of the
compaction_manager and subscribes to it. If the abort source is
triggered, the compaction manager will react to it firing and all
compactions it manages will be stopped.

We still want the database object to be able to wait for the compaction
manager, since the database is the object that owns the lifetime of
the compaction manager. To make that possible we'll use a future
that is return from stop(): no matter what triggered the abort, either
an early abort during initial resharding or a database-level event like
drain, everything will shut down in the right order.

The abort source is passed to the database, who is responsible from
constructing the compaction manager.

Tests: unit (debug), manual start+stop, manual drain + stop, previously
       failing dtests.

v5:
- after another revert, decided to clarify even further shutdown states.
  It is now explicitly legal to stop an object that was never started,
  once it is started it can be stop()d only once. Objects that were ever
  started must be stopped. This should do it (Famous last words)

v4:
- make do_stop public for the benefit of drain + shutdown.
- clarify role of future<> stop() vs void do_stop()
- fixed a call to stop to do_stop internally on compaction manager failure.
  We were ignoring the future anyway, so do_stop is the best there.
- removed a double stop() coming from scylla itself. It happened because
  main.cc had an explicit call to compaction_manager.stop(). That call
  was introduced in bf18025937 by Raphael with the goal of giving a
  head start to the compaction manager in stopping its compactions. This
  is now unnecessary, since this will be triggered by the abort source
  itself. In that sense, all this patch does is to move that ability a
  lot earlier and implement it over the abort_source which did not exist
  when Raphael added the stop calls.

Signed-off-by: Glauber Costa <glauber@scylladb.com>