distributed_loader.cc: remove old, unused do_populate_column_family
Population is now controlled by a new class,
directory_with_sstables_handler.  It works for any directory, including
upload, and guarantees all SSTables that are loaded are unique to a
shard. It supersedes do_populate_column_family in every way so now we
can remove it.

do_populate_column_family it's full of complications to handle shared
SSTables and we can now get rid of it.

Signed-off-by: Glauber Costa <glauber@scylladb.com>